moved logic to profile data hook
